Extract PM quality router models
